access of user to the card

+++ src/graphql/Card/Resolver.ts
@@ -13,27 +13,47 @@ const getOneCard = async (parent: any, args: any, context: Context) => {
   return existingCard
 }
 const createNewCard = async (parent: any, args: any, context: Context) => {
+  const userId = context.userId;
+  if (!userId) throw new Error("Access denied");
   const newCard = await context.prisma.card.create({
     data: {
       answer: args.answer,
-      question: args.question
+      question: args.question,
+      postedById: userId
     }
   })
   return newCard
 }
 
 const deleteCard = async (parent: any, args: any, context: Context) => {
-  const cardToBeDeleted = await context.prisma.card.delete({
+  const userId = context.userId;
+  if (!userId) throw new Error("Access denied");
+  const cardToBeDeleted = await context.prisma.card.findUnique({
     where: {
-      id: args.id,
+      id: args.id
     }
   })
-  if (!cardToBeDeleted) throw new Error("Card with that id is not found");
+  if (!cardToBeDeleted) throw new Error("card not found");
+  if (cardToBeDeleted.postedById !== userId) throw new Error("you have no access on this card");
+  await context.prisma.card.delete({
+    where: { id: args.id }
+  })
   return "card have been deleted"
 }
 
 const updateCard = async (parent: any, args: any, context: Context) => {
-  const cardToBeUpdated = await context.prisma.card.update({
+  const userId = context.userId;
+  if (!userId) throw new Error("Access denied");
+
+  const cardToBeUpdate = await context.prisma.card.findUnique({
+    where: {
+      id: args.id
+    }
+  })
+  if (!cardToBeUpdate) throw new Error("card not found");
+  if (cardToBeUpdate.postedById !== userId) throw new Error("you have no access on this card");
+
+  const updatedCard = await context.prisma.card.update({
     where: {
       id: args.id,
     },
@@ -42,15 +62,16 @@ const updateCard = async (parent: any, args: any, context: Context) => {
       answer: args.answer
     }
   })
-  if (!cardToBeUpdated) throw new Error("card with that id not found");
-  return cardToBeUpdated
+  return updatedCard
 
 }
-const postedByUser = async (parent: any, args: any, context: Context) => {
-  const cardfForUser = await context.prisma.user.findUnique({
-    where: { id: parent.id }
+const getOwnCards = async (parent: any, args: any, context: Context) => {
+  const userId = context.userId;
+  if (!userId) throw new Error("Access denied");
+  const ownCards = await context.prisma.card.findMany({
+    where: { postedById: userId }
   })
-  return cardfForUser
+  return ownCards
 }
 
-export { getAllCards, createNewCard, getOneCard, deleteCard, updateCard, postedByUser }
+export { getAllCards, createNewCard, getOneCard, deleteCard, updateCard, getOwnCards }
